# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Documentation for ITA-Wiki",
        "description": "Project ITA-Wiki documentation wall",
        "version": "1.0.0.0"
    },
    "paths": {
...
        "/api/resources/{resource}": {
            "put": {
                "tags": [
                    "Resources"
                ],
                "summary": "Update a resource",
                "description": "Update an existing resource with validation",
                "operationId": "2fb72cef844f47cbfbad3db9664d992c",
                "parameters": [
                    {
                        "name": "resource",
                        "in": "path",
                        "description": "ID of the resource to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "github_id",
                                    "title",
                                    "description",
                                    "url"
                                ],
                                "properties": {
                                    "github_id": {
                                        "type": "integer",
                                        "example": 6729608
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel Best Practices"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A collection of best practices for Laravel development"
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://laravelbestpractices.com"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "testing"
                                        },
                                        "maxItems": 5,
                                        "uniqueItems": true,
                                        "example": [
                                            "testing",
                                            "tdd",
                                            "hooks"
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You don't have permission to update this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required.",
                                                    "Title must be at least 5 characters.",
                                                    "Title must not exceed 255 characters."
                                                ],
                                                "description": [
                                                    "The description field is required.",
                                                    "Description must be at least 10 characters.",
                                                    "Description must not exceed 1000 characters."
                                                ],
                                                "url": [
                                                    "The url field is required.",
                                                    "The url format is invalid."
                                                ],
                                                "tags": [
                                                    "The tags field must be an array of strings.",
                                                    "The selected tags are invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Resource": {
                "title": "Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "github_id": {
                        "type": "integer",
                        "example": 12345
                    },
                    "title": {
                        "type": "string",
                        "example": "Lorem Ipsum ...",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Lorem Ipsum ...",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://www.hola.com",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "Node",
                            "React",
                            "Angular",
                            "JavaScript",
                            "Java",
                            "Fullstack PHP",
                            "Data Science",
                            "BBDD"
                        ],
                        "example": "Node"
                    },
                    "tags": {
                        "description": "Array of tags",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "kubernetes",
                            "sql",
                            "azure"
                        ]
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "Video",
                            "Cursos",
                            "Blog"
                        ],
                        "example": "Video"
                    },
                    "bookmark_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "like_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T19:23:41.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T19:23:41.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}